Refactor data-dense tables to mobile-first responsive lists
- Replaced traditional HTML <table> elements with Flexbox-based list rows in DeviceTable.tsx, VlanTable.tsx, and EventLog.tsx.
- Implemented a visual hierarchy for mobile: primary identifiers on the left, status/primary state on the right, and secondary details as subtext.
- Updated AGENTS.md to include Mobile-First UI/UX standards.
- Verified frontend build and resolved missing types/websocket.ts dependency.
- Updated E2E tests (tests/test_e2e_web_ui.py, tests/test_e2e_device_status.py) to align with the new DOM structure.
- Fixed broken relative imports in event/device and meraki_select modules and updated test imports to reference MerakiMainCoordinator.
