+++ custom_components/meraki_ha/meraki_select/vpn.py
@@ -10,7 +10,7 @@ from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.entity import EntityCategory
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from .coordinators import MerakiMainCoordinator
+from ..coordinators import MerakiMainCoordinator
 from ..core.api.client import MerakiAPIClient
 from ..core.models.network import MerakiNetwork, MerakiVpn
 from ..helpers.device_info_helpers import resolve_device_info
